# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: eco-base API
  version: "1.0.0"
  description: |
    Enterprise cloud-native AI platform API.
    URI: eco-base://backend/api/openapi

    Provides authentication, platform management, AI inference (OpenAI-compatible),
    vector alignment, async job management, embedding, and .qyaml governance.
  contact:
    name: eco-base Platform Team
    email: platform-team@autoecoops.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development (API gateway)
  - url: http://localhost:8001
    description: Local development (AI service)
  - url: https://api.autoecoops.io
    description: Production
...
paths:
...
  /api/v1/platforms:
    get:
      summary: List platforms
      tags: [Platforms]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform list
    post:
      summary: Register platform (admin only)
      tags: [Platforms]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
                capabilities: { type: array, items: { type: string } }
                deployTarget: { type: string }
      responses:
        "201":
          description: Platform created
        "403":
          description: Admin access required
  /api/v1/ai/generate:
    post:
      summary: Generate AI content (legacy)
      tags: [AI]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "504":
          description: AI service timeout
...
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat completions
      tags: [OpenAI]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Chat completion response (or SSE stream if stream=true)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                type: string
        "422":
          description: Validation error
...
  /v1/models:
    get:
      summary: OpenAI-compatible model list
      tags: [OpenAI]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
...
  /api/v1/yaml/generate:
    post:
      summary: Generate .qyaml governance manifest
      tags: [YAML]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [module]
              properties:
                module:
                  type: object
                  properties:
                    name: { type: string }
                    ports: { type: array, items: { type: integer } }
                    depends_on: { type: array, items: { type: string } }
                target: { type: string, default: "k8s" }
      responses:
        "200":
          description: Generated .qyaml content
  /api/v1/yaml/validate:
    post:
      summary: Validate .qyaml governance compliance
      tags: [YAML]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        "200":
          description: Validation result
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenerateRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string }
        model_id: { type: string, default: "default" }
        max_tokens: { type: integer, default: 2048 }
        temperature: { type: number, default: 0.7 }
    GenerateResponse:
      type: object
      properties:
        request_id: { type: string }
        content: { type: string }
        engine: { type: string }
        model_id: { type: string }
        uri: { type: string }
        urn: { type: string }
        usage: { type: object }
        latency_ms: { type: number }
    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        model: { type: string, default: "default" }
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role: { type: string, enum: [system, user, assistant] }
              content: { type: string }
        temperature: { type: number, minimum: 0, maximum: 2 }
        max_tokens: { type: integer, default: 2048 }
        stream: { type: boolean, default: false }
    ChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, example: chat.completion }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message:
                type: object
                properties:
                  role: { type: string }
                  content: { type: string }
              finish_reason: { type: string }
        usage:
          $ref: "#/components/schemas/UsageInfo"
        system_fingerprint: { type: string }
...
    ModelListResponse:
      type: object
      properties:
        object: { type: string, example: list }
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              object: { type: string, example: model }
              owned_by: { type: string, example: eco-base }
              capabilities: { type: array, items: { type: string } }
              compatible_engines: { type: array, items: { type: string } }
              context_length: { type: integer }
...
    UsageInfo:
      type: object
      properties:
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }